Project/Bot/requirements.txt
flask
openai
smartapi-python
logzero
websocket-client
